[ISSUE #9689]♻️Extract Tokio server launch and connection owners
